requirements.txt
pandas>=2.1
numpy>=1.26
pyarrow>=15.0
PyYAML>=6.0

wrds
